+++ src/common/utils/workaroundUtils.ts
@@ -5,6 +5,7 @@ import { createTempDir, uxLog } from './index.js';
 import { glob } from 'glob';
 import { parseXmlFile, writeXmlFile } from './xmlUtils.js';
 import { isScratchOrg } from './orgUtils.js';
+import { GLOB_IGNORE_PATTERNS } from './projectUtils.js';
 
 // Update files for special cases
 export async function arrangeFilesBefore(commandThis: any, options: any = {}) {
@@ -21,7 +22,7 @@ export async function arrangeFilesBefore(commandThis: any, options: any = {}) {
 export async function removeLookupFilters(tempDir: string, commandThis: any, options: any = {}) {
   const arrangedFiles: any = [];
   const findFieldsPattern = (options.rootFolder || '.') + `/**/objects/**/fields/**.field-meta.xml`;
-  const matchingFieldFiles = await glob(findFieldsPattern, { cwd: process.cwd() });
+  const matchingFieldFiles = await glob(findFieldsPattern, { cwd: process.cwd(), ignore: GLOB_IGNORE_PATTERNS });
   for (const fieldFile of matchingFieldFiles) {
     // skip if managed field
     if ((path.basename(fieldFile).match(/__/g) || []).length === 2) {