+++ CHANGELOG.md
@@ -4,9 +4,18 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- New AI Provider: Agentforce
+- Create Objects AI-powered documentation
+  - Summary
+  - Relationships with other objects
+  - Fields
+  - Validation rules
+  - Related flows
+- Handle prompts multilingualism (ex: `PROMPTS_LANGUAGE=fr`)
+- Handle prompts cache to save tokens
 - Add `SFDX_DISABLE_FLOW_DIFF: false` in default CI/CD pipelines (must be set to true during CI/CD setup)
 - Enhance branches & orgs CI/CD strategy mermaid diagram
-- Handle prompts multilingualism and cache
+- Improve performances by using `GLOB_IGNORE_PATTERNS` for all calls to glob
 
 ## [5.15.5] 2025-01-16